Add user validations and protect password & salt from being leaked.
The password and salt are protected by preventing code from accessing password and salt except before creation. This is done by setting the default scope for the user model to one that excludes the password and salt.
We manually reload the model in an afterCreate hook to ensure that the new defaultScope is applied to the returned object.

More information in this StackOverflow post and in the comments: https://stackoverflow.com/a/48357983/7056420