Fix issues with seeds not running properly.
Implementation details:
As per https://github.com/knex/knex/issues/2010, knex intentionally runs seeds every time the application starts so it's up to us to define the logic to re-run seeds.
Hence, we add a check to see if the seed data is present (check if the seeded users are present). If so, don't run seeds. Otherwise, delete the seed data in case it was modified, and re-run seeds.