Make login secure (at the expense of user experience).
The authentication token is now stored in memory instead of local storage. This means login is now perfectly secure, but page refreshes cause the user to lose their session.

In the future, we will want to persist sessions with a refresh token that can be securely stored in localstorage.

We will also probably use something other than the apollo client in-memory cache (eg. react context api) because it doesn't look like it's intended to be read directly as I am doing.

Despite those drawbacks, I prefer putting security above usability until I have time to implement the best of both worlds solution, as outlined in the comment I left in useLogin.js.