Add frontend kubernetes configuration.
Also fix wrong container ports exposed in dockerfiles.

Note: the frontend website currently cannot contact the backend when running in the cluster. This is because the external IP is determined by minikube after the configurations are applied. More research will be needed to figure out the best way to get this working locally.