Fix "some backend services are in an UNHEALTHY state" error on GCP.
It was in part caused by the fact that the "/" path redirects to "/login". A redirect returns a 304 http status code, but the liveness probe must return 200 so we specify a different url.
To address this, make the root url show the login page (when not logged in), and redirect to "/repositories" when logged in.

One other issue was that the k80s frontend-service had the wrong containerPort.

This commit also ensures that the latest images are always used when deploying with helm and ensures that the k80s cluster is at the most recent version.