Fix issue where environment variables were not accessible on the frontend for prod builds.
In short, we create our own server for prod where we inject environment variables in the html.

For prod, combination of the following two strategies:
https://create-react-app.dev/docs/title-and-meta-tags/#injecting-data-from-the-server-into-the-page
https://create-react-app.dev/docs/deployment/#other-solutions

In dev mode, we use environment variables prefixed with 'REACT_APP_' (which doesn't work with production builds)
More info: https://create-react-app.dev/docs/adding-custom-environment-variables#referencing-environment-variables-in-the-html

In the code, these two approaches are merged in the src/utils/environment.js file so that we can just import the variables.
In short, there are three relevant files: server/app.js, public/index.html, and src/utils/environment.js