Remove issue url normalization on insert.
Since we can't automatically normalize the url on all find queries, it doesn't make sense to normalize the url in the database. As per https://github.com/Vincit/objection.js/issues/603, there isn't really a where/findOne hook. The beforeFind hook is also not useful in this case because its StaticHookArguments are empty (ie. they don't include the arguments passed in "find").

The alternative would be to call normalizeUrl manually everywhere, but it's very error-prone because forgetting a single one would mean that we can't reliably search by issue url.

At the moment, the better tradeoff is to simply accept the risk that someone may refer to the same issue twice with slightly different urls, and will therefore get two tasks created when the issue is closed. A minor inconvenience.