Change eslint errors to warnings.
Set max lint warnings to 0 in CI.

This fixes the issue in react-scripts 4.0.0 where eslint errors prevent the app from running. The desired behaviour is for the app to run if it can in development because we will often comment-out blocks of code or add console.logs, and we don't want that to prevent running the app. CI will ensure that those lint warnings from dev changes don't make their way into production.