Set initial delay for backend probes to 10 minutes
This is to see if the issue is just that we don't wait long enough for the backend to be available.

I don't think the application startup is the problem since we can see in the logs that the app does start in time. Maybe there is something with ports being open later however.